Extracts text reference update methods
